CAMS-49: renamed http.humble.ts to http.adapter.ts
