+++ user-interface/src/components/utils/http.adapter.ts
